Add Fedora 22 repository
Add Fedora 22 repository and remove Fedora 18 and 19 repositories.
Technically the older repositories still exist but they're officially
unsupported.

Use 'yum install' rather than 'yum localinstall' because it also
works for now with dnf.  This documentation should be updated for
Fedora 22 to reflect the new package manager but for now this works.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>