Add in zfs-release dir with spec files (#68)
In the past we've simply checked in the zfs-release RPM files
directly.   It was requested that we start checking in the
spec and repo files too so that you can easily see the diff
when making changes.  This commit checks in all the files
needed to create the RPMs, along with instructions on how
to build them.

Signed-off-by: Tony Hutter <hutter2@llnl.gov>
Closes #63