Added stream10 to matrix runners, added EL10 to zfs-release.spec (#96)
CentOS Stream10 and other EL10 releases need a build a repo available for
them. Used the new GPG key and symlinked for building EL10 releases.
Ensured that stream10 is part of the build matrix and and stream 10 to
the list of runners. Note: this will increase build times as it is
adding a 4th build to be run in addtion to AlmaLinux9, F41, and F40.

Signed-off-by: James Reilly <jreilly1821@gmail.com>
Reviewed-by: Tony Hutter <hutter2@llnl.gov>